Fix Linux build: swr_convert const cast, int64_t %lld format
- preview-encoder.cpp: swr_convert() takes 'const uint8_t **', not
  'const uint8_t* const*'. Add const_cast at the call site so GCC's
  strict overload resolution accepts the conversion.

- buffer-manager.cpp: on x86_64 Linux int64_t is 'long', not 'long
  long', so '%lld' against MAX_DRIFT_FOR_CATCHUP_MS / drift_from_live_ms
  triggers -Wformat. Cast both args to 'long long' to match the format
  specifier across platforms.

+++ src/utils/buffer-manager.cpp
@@ -80,7 +80,8 @@ bool BufferManager::should_skip_to_live() const
 	if (should_skip) {
 		lss_log_warn("BufferManager: Drift too large for catch-up, should skip to live "
 		             "(drift=%lld ms, threshold=%lld ms)",
-		             state_.drift_from_live_ms, MAX_DRIFT_FOR_CATCHUP_MS);
+		             (long long)state_.drift_from_live_ms,
+		             (long long)MAX_DRIFT_FOR_CATCHUP_MS);
 	}
 
 	return should_skip;
